Refactor assignment and subject models, update services
Refactored assignment and subject models to include navigation fields (termId, subjectId, subjectName) extracted from Firestore document paths. Updated assignment status handling to use booleans instead of strings throughout repositories and viewmodels. Removed SubjectService and migrated subject/weight logic to repository pattern. Adjusted assignment creation and editing to use weightId and new weight selection UI. Updated various viewmodels and screens to align with these changes, and improved holiday date handling.

+++ lib/services/grades/gpa_calculation_service.dart
@@ -36,7 +36,7 @@ class GpaCalculationService {
       // Note: In a real implementation, you'd parse weightCategoriesJson
       // For now, calculate simple weighted average
       double totalWeightedGrade = 0.0;
-      int totalWeight = 0;
+      double totalWeight = 0.0;
 
       for (final assignment in gradedAssignments) {
         totalWeightedGrade += assignment.grade * assignment.weight;
